image: move image identity logic to dedicated file
